moves spinner into the button vs the entire card
